requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
colorclass==2.2.0
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.0
docopt==0.6.2
filelock==3.0.12
google==3.0.0
google-api-core==1.26.2
google-auth==1.28.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.5
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
pbr==5.5.1
Pillow==8.1.2
pip-upgrader==1.4.15
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.7.2
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
stevedore==3.3.0
terminaltables==3.1.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
wrapt==1.12.1
